재즈, 팝 문항 추가
