+++ music/src/Pages/SecondTest/SecondPage4.js
@@ -3,123 +3,46 @@ import { Link } from "react-router-dom";
 import { ChakraProvider } from "@chakra-ui/react";
 import { Progress } from "@chakra-ui/progress";
 
-import RbBox from "../../Components/RbBox";
-import { PopValueState } from "../../Components/PopBox";
-import { HipValueState } from "../../Components/HipBox";
-import { JazzValueState } from "../../Components/JazzBox";
-import { RbValueState } from "../../Components/RbBox";
-import { RockValueState} from "../../Components/RockBox";
-import { useRecoilValue } from "recoil";
+import PopBox from "../../Components/PopBox";
 
 import "./SecondPage.css";
 
-function combineKeysValues(obj) {
-    const keys = Object.keys(obj);
-    const values = Object.values(obj);
-    const combined = {};
-
-    for (let i = 0; i < keys.length; i++) {
-        combined[keys[i]] = values[i];
-    }
-
-    return combined;
-}
 const SecondPage4 = () => {
-    const PopCheckValue = useRecoilValue(PopValueState); 
-    const HipCheckValue = useRecoilValue(HipValueState);
-    const JazzCheckValue = useRecoilValue(JazzValueState);
-    const RbCheckValue = useRecoilValue(RbValueState);
-    const RockCheckValue = useRecoilValue(RockValueState);
-
-    // 각 장르당 합 구하기 
-    const PopTotal = Object.values(PopCheckValue).reduce((sum,value)=> sum + value, 0);
-    const HipTotal = Object.values(HipCheckValue).reduce((sum,value)=> sum + value, 0);
-    const JazzTotal = Object.values(JazzCheckValue).reduce((sum,value)=> sum + value, 0);
-    const RbTotal = Object.values(RbCheckValue).reduce((sum,value)=> sum + value, 0);
-    const RockTotal = Object.values(RockCheckValue).reduce((sum,value)=> sum + value, 0);
-
-    const AllTotal = {
-        Pop: PopTotal,
-        Hip: HipTotal,
-        Jazz: JazzTotal,
-        Rb: RbTotal,
-        Rock: RockTotal,
-    };
-
-    const Result = Object.keys(AllTotal).reduce((max, genre) => {  // 최대값 가진 장르
-        return AllTotal[genre] > AllTotal[max] ? genre : max;
-    }, 'Pop')
-
-    const maxTotal = AllTotal[Result]; // 최대값
-
-    const completeButton = () => {
-        console.log("결과 :" , Result);
-        console.log("장르 :", maxTotal);
-        sendDataToServer();
-    };
-    const sendDataToServer = async () => {
-        const data = {
-            pop: combineKeysValues(PopCheckValue),
-            hip: combineKeysValues(HipCheckValue),
-            jazz: combineKeysValues(JazzCheckValue),
-            rb: combineKeysValues(RbCheckValue),
-            rock: combineKeysValues(RockCheckValue)
-        }
-        try {
-          const response = await fetch("https://34.64.108.76.nip.io/answer", {
-            method: 'POST',
-            headers: {
-              'Content-Type': 'application/json'
-            },
-            body: JSON.stringify(data)
-          });
-      
-          if (!response.ok) {
-            throw new Error('Network response was not ok');
-          }
-      
-          const responseData = await response.json();
-          console.log('Server response:', responseData);
-        } catch (error) {
-          console.error('Error sending data:', error);
-        }
-      };      
 
     return (
         <ChakraProvider>
         <div className="firstPage">
             <div className="progress">
-                <Progress value={100} size='md' colorScheme='purple' />
+                <Progress value={56} size='md' colorScheme='purple' />
             </div>
             <div className="num2">
-                <h3>디지털한 사운드가 기반이 된 끈적한 사랑과 사회참여적인 가사를 담아낸 흑인들의 가사
-그리고 고요하고 그루비한 보컬..<br/> 오우.. 이런 노래엔 스모키한 위스키 한잔이 딱이야.</h3>
-                <RbBox id="NeoRB1"/>
+                <h3>긍정적인 에너지를 뿜어내는 업템포 음악과 단순하면서 중독성이 강한 멜로디는 내 어깨를 들썩이게 해!</h3>
+                <PopBox id="DancePop2"/>
             </div>
             <div className="num2">
-                <h3>현대적인 다양한 장르를 결합해 트렌드를 이끄는 사운드 위에 보컬의 소울과 그루브가 느껴져.
-이런 세련된 섹시함은 나의 미간을 찡그리게 만들지.</h3>
-                <RbBox id="ComRB1"/>
+                <h3>전자 악기와 디지털한 사운드의 조화!
+마치 다채롭고 현란한 레이저쇼 현장에 온 기분이야!</h3>
+                <PopBox id="ElecPop2"/>
             </div>
             <div className="num2">
-                <h3>끈적하면서 그루비한 보컬의 소울과 드럼의 리듬 위에 째지한 사운드.. 
-독특하면서 센스있는 베이스의 그루브는 나를 춤추게 만들지 I Feel Good~</h3>
-                <RbBox id="PunkRB1"/>
+                <h3>전반적으로 밝고 경쾌하지만 가사에는 우울함과 현실의 분노가 담긴 음악.. 
+                    마치 나의 양면적인 모습과 닮았어.</h3>
+                <PopBox id="BritPop2"/>
             </div>
             <div className="num2">
-                <h3>교회에서 들릴듯한 전통적인 구성들과 흑인민권 운동을 기반으로 흑인의 한과 울림이 담긴 음악.<br/>
-무대 옆에 있는 코러스(콰이어)단 들이 하모니를 만들때 난 그시대를 입체적으로 떠올리게 해.</h3>
-                <RbBox id="SoulRB1"/>
+                <h3>분위기 있는 작은 공연장에 나만 알고싶은 아티스트가 콘서트를 열었어.
+감성적인 멜로디와 차분한 선율..이게 나의 안정제지.</h3><br/>
+                <PopBox id="IndiPop2"/>
             </div>
             <div className="num2">
-                <h3>자신만의 아이덴티티로 창의성을 펼쳐내 복잡한 자신의 감정을 다양하고 특이한 사운드로 표현하는
-소울이 담겨진 음악.. 이건 아트가 아닐까?</h3>
-                <RbBox id="AlterRB1"/>
+                <h3>스페인어만의 특유한 어감과 레게톤의 음악!
+마치 스페인 바닷가에서 개최한 뮤직 페스티벌에 와 있는것 같아서 정말 신나!</h3><br/>
+                <PopBox id="LatinPop2"/>
             </div>
             <div className='linkBox'>
-                <Link to={`/${Result}`}>
-                    <button onClick={completeButton}>완료</button> 
-                </Link>
+                <Link to="/SecondPage5">
+                    <button>다음으로</button> 
+                </Link> 
             </div>
         </div>
         </ChakraProvider>
